fix search for meetings

+++ protected/controllers/CategoryController.php
@@ -111,12 +111,18 @@ class CategoryController extends Controller
 	{
 		if(Yii::app()->request->isPostRequest)
 		{
-			// we only allow deletion via POST request
-			$this->loadModel($id)->delete();
+			$meeting = Meeting::model()->find('category_id = :id', array(':id'=>$id,));
+			if($meeting) {
+				throw new CHttpException(400,'Invalid request. Please dellete mettings with this category.');
+			}
+			else {
+				// we only allow deletion via POST request
+				$this->loadModel($id)->delete();
 
-			// if AJAX request (triggered by deletion via admin grid view), we should not redirect the browser
-			if(!isset($_GET['ajax']))
-				$this->redirect(isset($_POST['returnUrl']) ? $_POST['returnUrl'] : array('admin'));
+				// if AJAX request (triggered by deletion via admin grid view), we should not redirect the browser
+				if(!isset($_GET['ajax']))
+					$this->redirect(isset($_POST['returnUrl']) ? $_POST['returnUrl'] : array('admin'));
+			}
 		}
 		else
 			throw new CHttpException(400,'Invalid request. Please do not repeat this request again.');
